sys-apps/entropy: fix repositories.conf relocation
git-svn-id: http://svn.sabayonlinux.org/overlay@2782 d7aec97c-591d-0410-af39-a8856400b30a

+++ sys-apps/entropy/entropy-0.51.0.ebuild
@@ -80,9 +80,20 @@ src_install() {
 
 }
 
+pkg_preinst() {
+	# backup user repositories.conf
+	if [ -f "${REPO_CONFPATH}" ]; then
+		cp -p "${REPO_CONFPATH}" "${REPO_CONFPATH}.backup"
+	fi
+}
+
 pkg_postinst() {
 	# Copy config file over
-	if [ -f "${REPO_CONFPATH}.example" ] && [ ! -f "${REPO_CONFPATH}" ]; then
-		cp ${REPO_CONFPATH}.example ${REPO_CONFPATH} -p
+	if [ -f "${REPO_CONFPATH}.backup" ]; then
+		cp ${REPO_CONFPATH}.backup ${REPO_CONFPATH} -p
+	else
+		if [ -f "${REPO_CONFPATH}.example" ] && [ ! -f "${REPO_CONFPATH}" ]; then
+			cp ${REPO_CONFPATH}.example ${REPO_CONFPATH} -p
+		fi
 	fi
 }
